backend/requirements.txt
fastapi
uvicorn
python-multipart
biopython
numpy
elasticsearch>=8.0.0,<9.0.0openai
